23.12.06 별점 img 추가전 백업커밋

+++ src/main/java/C_Team/MovieStar/dto/StarDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 public class StarDto {
     public String userId;
-    public int starCount;
+    public float starCount;
     public int movieCode;
 }
